morph_mesh_scene / morph_targets_scene: headerless
